backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
slowapi==0.1.9
httpx==0.27.2
# Pydantic v2 with pre-built wheels for Python 3.12+
pydantic==2.10.6
# Supabase
supabase==2.10.0
